+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/config/SecurityConfig.java
@@ -54,6 +54,7 @@ public class SecurityConfig {
                             "/css/**",
                             "/swagger-ui/**",
                             "/v3/api-docs/**").permitAll()
+                        .requestMatchers("/api/admin/**").hasRole("ADMIN")
                         .anyRequest().authenticated()
                 )
                 .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);