+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/request/DoctorCreateDTO.java
@@ -1,5 +1,6 @@
 package com.ivanovp.medical_record.dto.request;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
@@ -23,6 +24,7 @@ public class DoctorCreateDTO {
     @NotNull
     private Long specialtyId;
 
+    @JsonProperty("isGp")
     private boolean isGp;
 
     @NotBlank