+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/response/DoctorResponseDTO.java
@@ -1,5 +1,6 @@
 package com.ivanovp.medical_record.dto.response;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,5 +14,6 @@ public class DoctorResponseDTO {
     private String uin;
     private String name;
     private String specialtyName;
+    @JsonProperty("isGp")
     private boolean isGp;
 }