+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/response/PatientResponseDTO.java
@@ -1,5 +1,6 @@
 package com.ivanovp.medical_record.dto.response;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,6 +13,7 @@ public class PatientResponseDTO {
     private Long id;
     private String name;
     private String egn;
+    @JsonProperty("isInsured")
     private boolean isInsured;
     private String gpName;
 }